GLUE-245 Refactor: 폴더구조 변경, appliction.yml 활성화
