+++ src/main/java/com/justdo/plug/auth/MemberApplication.java
@@ -1,0 +1,19 @@
+package com.justdo.plug.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@SpringBootApplication
+@EnableFeignClients(basePackages = {"com.justdo.plug"})
+@EnableDiscoveryClient
+public class MemberApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MemberApplication.class, args);
+	}
+
+}
