+++ src/main/java/com/justdo/plug/auth/domain/member/dto/request/MemberInfoRequest.java
@@ -1,0 +1,21 @@
+package com.justdo.plug.auth.domain.member.dto.request;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+public class MemberInfoRequest {
+    private String email;
+    private String nickname;
+    private String profile_url;
+
+    @Builder
+    public MemberInfoRequest(String email, String nickname, String profile_url) {
+        this.email = email;
+        this.nickname = nickname;
+        this.profile_url = profile_url;
+    }
+}
