+++ src/main/java/com/justdo/plug/auth/global/response/code/BaseCode.java
@@ -1,0 +1,7 @@
+package com.justdo.plug.auth.global.response.code;
+
+public interface BaseCode {
+    public ReasonDto getReason();
+
+    public ReasonDto getReasonHttpStatus();
+}
