+++ src/main/java/com/justdo/plug/auth/global/response/code/BaseErrorCode.java
@@ -1,0 +1,7 @@
+package com.justdo.plug.auth.global.response.code;
+
+public interface BaseErrorCode {
+    public ErrorReasonDto getReason();
+
+    public ErrorReasonDto getReasonHttpStatus();
+}
