+++ src/main/java/com/justdo/plug/auth/global/response/code/ReasonDto.java
@@ -1,0 +1,19 @@
+package com.justdo.plug.auth.global.response.code;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ReasonDto {
+
+    private HttpStatus httpStatus;
+    private final boolean isSuccess;
+    private final String code;
+    private final String message;
+
+    public boolean getIsSuccess() {
+        return isSuccess;
+    }
+}
