+++ src/main/java/com/justdo/plug/auth/global/utils/redis/RedisUtils.java
@@ -1,0 +1,28 @@
+package com.justdo.plug.auth.global.utils.redis;
+
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.stereotype.Component;
+
+import java.util.concurrent.TimeUnit;
+
+@Component
+public class RedisUtils {
+
+    private final RedisTemplate<String, Object> redisTemplate;
+
+    public RedisUtils(RedisTemplate<String, Object> redisTemplate) {
+        this.redisTemplate = redisTemplate;
+    }
+
+    public void setData(String key, String value,Long expiredTime){
+        redisTemplate.opsForValue().set(key, value, expiredTime, TimeUnit.MILLISECONDS);
+    }
+
+    public String getData(String key){
+        return (String) redisTemplate.opsForValue().get(key);
+    }
+
+    public void deleteData(String key){
+        redisTemplate.delete(key);
+    }
+}
