--- src/main/java/com/justdo/plug/blog/BlogClient.java
@@ -1,11 +1,0 @@
-package com.justdo.plug.blog;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@FeignClient(name = "blog-service",url="${application.config.blogs-url}")
-public interface BlogClient {
-    @PostMapping
-    void createBlog(@RequestParam Long memberId);
-}
