--- src/main/java/com/justdo/plug/member/global/config/SecurityConfig.java
@@ -1,53 +1,0 @@
-package com.justdo.plug.member.global.config;
-
-import com.justdo.plug.member.global.jwt.kakao.CustomAuthenticationSuccessHandler;
-import com.justdo.plug.member.global.jwt.kakao.OAuth2UserService;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.CorsConfigurationSource;
-
-import java.util.Arrays;
-import java.util.Collections;
-
-@Configuration
-@EnableMethodSecurity
-@RequiredArgsConstructor
-public class SecurityConfig {
-    private final OAuth2UserService oAuth2UserService;
-    private final CustomAuthenticationSuccessHandler successHandler;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.csrf(AbstractHttpConfigurer::disable);
-        http.cors(corsCustomizer -> corsCustomizer.configurationSource(new CorsConfigurationSource() {
-            @Override
-            public CorsConfiguration getCorsConfiguration(HttpServletRequest request) {
-                CorsConfiguration config = new CorsConfiguration();
-                config.setAllowedOrigins(Collections.singletonList("http://front_url"));
-                config.setAllowedMethods(Collections.singletonList("*"));
-                config.setAllowCredentials(true);
-                config.setAllowedHeaders(Collections.singletonList("*"));
-                config.setMaxAge(3600L);
-                config.setExposedHeaders(Arrays.asList("Authorization", "Authorization-refresh"));
-                return config;
-            }
-        }));
-        http.authorizeHttpRequests(config -> config.anyRequest().permitAll());
-
-        http.oauth2Login((oauth2) -> oauth2
-                .loginPage("/login")
-                .successHandler(successHandler)
-                .userInfoEndpoint(userInfoEndPoint -> userInfoEndPoint.userService(oAuth2UserService))
-        );
-
-        return http.build();
-    }
-
-}
