--- src/main/java/com/justdo/plug/member/global/config/SwaggerConfig.java
@@ -1,45 +1,0 @@
-package com.justdo.plug.member.global.config;
-
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import io.swagger.v3.oas.models.servers.Server;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-
-        String jwtSchemeName = "JWT TOKEN";
-
-        // API 요청 헤더에 인증정보 포함
-        SecurityRequirement securityRequirement = new SecurityRequirement().addList(jwtSchemeName);
-
-        //SecuritySchemes 등록
-        Components components = new Components()
-            .addSecuritySchemes(jwtSchemeName, new SecurityScheme()
-                .name(jwtSchemeName)
-                .type(SecurityScheme.Type.HTTP)
-                .scheme("bearer")
-                .bearerFormat("JWT"));
-
-        return new OpenAPI()
-            .addServersItem(new Server().url("/"))
-            .info(apiInfo())
-            .addSecurityItem(securityRequirement)
-            .components(components);
-    }
-
-    private Info apiInfo() {
-        return new Info()
-            .title("Glue Auth-Service Springdoc 테스트")
-            .description("Springdoc을 사용한 Glue Auth-Service UI 테스트")
-            .version("1.0.0");
-    }
-}
