--- src/main/java/com/justdo/plug/member/global/jwt/kakao/OAuth2UserService.java
@@ -1,28 +1,0 @@
-package com.justdo.plug.member.global.jwt.kakao;
-
-import jakarta.transaction.Transactional;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-
-@Service
-@Transactional
-public class OAuth2UserService extends DefaultOAuth2UserService {
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        OAuth2User oAuth2User = super.loadUser(userRequest);
-
-        String userNameAttributeName = userRequest.getClientRegistration()
-                .getProviderDetails()
-                .getUserInfoEndpoint()
-                .getUserNameAttributeName();
-
-        return new DefaultOAuth2User(oAuth2User.getAuthorities(), oAuth2User.getAttributes(), userNameAttributeName);
-    }
-
-}
-
