--- src/main/java/com/justdo/plug/member/global/response/code/BaseCode.java
@@ -1,7 +1,0 @@
-package com.justdo.plug.member.global.response.code;
-
-public interface BaseCode {
-    public ReasonDto getReason();
-
-    public ReasonDto getReasonHttpStatus();
-}
