--- src/main/java/com/justdo/plug/member/global/response/code/BaseErrorCode.java
@@ -1,7 +1,0 @@
-package com.justdo.plug.member.global.response.code;
-
-public interface BaseErrorCode {
-    public ErrorReasonDto getReason();
-
-    public ErrorReasonDto getReasonHttpStatus();
-}
