+++ src/main/resources/application.yml
@@ -1,0 +1,7 @@
+spring:
+  application:
+    name: auth
+  profiles:
+    active: dev
+  config:
+    import: optional:configserver:http://localhost:8888
