+++ src/test/java/com/justdo/plug/auth/MemberApplicationTests.java
@@ -1,0 +1,13 @@
+package com.justdo.plug.auth;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MemberApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
