--- src/test/java/com/justdo/plug/member/MemberApplicationTests.java
@@ -1,13 +1,0 @@
-package com.justdo.plug.member;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MemberApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
